requirements.txt
certifi==2022.12.7
charset-normalizer==2.1.1
colorama==0.4.6
filelock==3.9.0
fsspec==2023.4.0
idna==3.4
Jinja2==3.1.2
llvmlite==0.41.1
MarkupSafe==2.1.3
more-itertools==10.2.0
mpmath==1.3.0
networkx==3.0
numba==0.58.1
numpy==1.24.1
openai-whisper==20231117
Pillow==9.3.0
PyAudio==0.2.14
regex==2023.12.25
requests==2.28.1
sympy==1.12
tiktoken==0.5.2















torch==2.1.2+cu121
torchaudio==2.1.2+cu121
torchvision==0.16.2+cu121
tqdm==4.66.1
typing_extensions==4.4.0
urllib3==1.26.13
Wave==0.0.2
